fix OnHTML function in scrape function
There was a bug in scrape function where OnHTML was called.
The calling function used var `htmlEl` which was not (correctly) available inside the function.
Only the last assigned value was used here, resulting in wrong map keys in the return.
Because the selector is not directly available in the HTMLCallback function defined in OnHTML the goquery selector has to be recreated.